# improved draw_text_mut performance