@theotherphil 2024-06-07
Thanks!

(Aside: text drawing regression tests https://github.com/image-rs/imageproc/issues/244 would be really useful to help avoid breakages from refactoring in this area!)